Create AlternatingCharacters.java
Given a string containing characters A and B only, we want to change it into an alternating string like ABABA. To do this, the characters in the string can be deleted. 
Find the minimum number of required deletions.

Input Format

The first line contains an integer T, i.e. the number of test cases. 
The next T lines contain a string each.

Output Format

For each test case, print the minimum number of deletions required.